requirements.txt
alembic==0.8.10
appdirs==1.4.0
click==6.7
Flask==0.12
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==0.23
packaging==16.8
pycrypto==2.6.1
PyJWT==1.4.2
PyMySQL==0.7.10
pyparsing==2.1.10
python-editor==1.0.3
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.5
Werkzeug==0.11.15
